rename customer viewing command
